Version 0.6.1

+++ src/main/java/com/knightsheraldry/items/item/khweapon/Billhook.java
@@ -31,9 +31,9 @@ public class Billhook extends SwordItem {
                     .forEach(entity -> {
                         boolean critical = false;
                         double distance = playerPos.distanceTo(target.getPos());
-                        float damage = SCDamageCalculator.getSCDamage(playerEntity, SCWeaponUtil.calculateDamage(this, distance,
+                        double damage = SCDamageCalculator.getSCDamage(playerEntity, SCWeaponUtil.calculateDamage(this, distance,
                                 damageType), damageType);
-                        float maxDamage = SCWeaponUtil.getDamageValues(damageType, this);
+                        double maxDamage = SCWeaponUtil.getDamageValues(damageType, this);
 
                         if (damage >= maxDamage) {
                             critical = true;
